requirements.txt
pandas==2.3.3
numpy==2.3.5
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
imbalanced-learn==0.13.0
xgboost==2.1.4
optuna==4.2.1
streamlit==1.42.2
shap==0.46.0
sqlalchemy==2.0.38
joblib==1.4.2
fpdf2==2.8.2
plotly==6.0.0
pytest==8.3.4
